Allows you to see other players shooting
On your own client you can now see the gunshots of the other players
currently connected to the server. this feature isn't perfect however
because the server just sends all gunshot data to connected clients even
if the given gunshot data was already processed so it may have duplicate
gunshots.