Fixed drawing glitch with bullets on clients
on clients bullets would look as if they passed through other objects
when in fact they had collided. this is because the bullet collision
detection was being run in the physics loop which may not run sometimes
due to lack of accumulated time, thus skipping the bullet collision
detection. bullets now check for collision in the time independent
object's update loop