Flag Position Bug Fix
If a teams flag is not being held by a player and its not located at the
teams base then it's position is unknown. if a new player connects they
will see the flag at the teams base because the server doesn't send
position updates for the flag. this is fixed so the server sends
position updates for the flags so a flag's position is always known.