Added a floor system
There is now a floor system where every entity is designated a floor to
exist on, ex. overground, underground.

objects can only interact with other objects if they are in the same
floor. in addition clients can only see objects that are on the same
floor as them. minimaps however will only show the overground floor, and
never the underground floor.