Added class to handle large Images
Created a class that loads and displays large images.
because sf::Texture has a size limit dependant on the graphics card you
have to split large images into smaller chunks and draw those instead.

currently the class only loads and draws an image, inaddition the top
left of the image is always at (0, 0) and cannot be moved.