Added a terrible Spatial Hash for blocks
The floor is divided into smaller sections and the blocks are placed
into the smaller sections. when checking for collision the entity only
checks for collision with blocks in the same section as the entity.

When split into 4 sections collision detection is 2x as fast as with no
splitting. currently there are 8 sections however adding the extra
sections don't seem to have any effect on the speed.