Server sends player's scores to clients, BUG
The server now sends the player's kill, death, capture, and return count
to all clients so that everyone can see the other player's scores.

***Bug
for a short while, after the server began sending the kill count, the
data received on the client side regarding the information about other
players would arrive corrupted.
when the data was read, the health for example would be the max value
for an unsigned integer. This was true for all other data being sent,
and as a result the number of bullets fired would be equal to the max
size of an unsigned int and the client would attempt to create thousands
of bullets and crash due to lack of memory.

This bug dissapered due to unknown reasons. i have no idea what caused
this bug, or why it disappeared. The bug dissapeared after i restarted
my router, however i have no idea if the problem was due to my code or
some external thing such as my router.