H.U.D displays Ammo, Held flag doesn't draw
The heads up display now displays the current ammo count for the player,
the total ammo in each magazine, and the total ammo for the entire gun.

Flags no longer draw when held by players. It is up to the players to
draw an icon to indicate he is holding a flag.