Added test bushes, increased minimap size
Added test bushes for players to hide behind and to check if hiding
mechanic is working properly. So far it works - opposing team members
hiding behind bushes aren't seen unless you're in the bush with them or
your teammate is in the bush.

Increased the size of the minimap so players can see more on the map.