fixed level editor object mirroring bug
Objects are mirrored horizontally along the center of the level, and the
center of the level is dependent on the size of the level. Level size is
dependent on the background image but the level size is only changed
when a background is set on the level, not when a level is loaded with a
background.

The bug was the level size would be 0 when the level is loaded because
the level size is not set unless the user sets a new background, thus
causing objects to mirror along an incorrect axis if the user loads a
level without setting a new background.

this is fixed so that the level size is checked everytime a mirrored
object is created so that the reflection axis is always correct.