Smoothened interpolation for other players
Other player's connected to the game now look as if they're smoothly
moving. This is because the rate at which the client sends position
updates, and the rate at which the server sends client updates to other
clients are now synced up.

Also once an object interpolates it sets whatever attribute its
interpolating to the target attribute.