Bug fixes: bullet drawing, portal use
Bug: bullets draw even if they're on a different floor.

this has been fixed so bullets only draw if they are on the same floor
as the player.

bug: when using portals to go between floors sometimes it makes you
stuck in a wall

this has been fixed. the problem was that after the server changed the
player's floor with the portal, the client would send a position update
before the server could send a floor update so the player on the server
side would be moved to the client's position which would be inside a
wall.

now the server ignores client position updates if the client's floor is
not the same as the server's floor.