probably fixed some lag and jitter
when connecting to other players over a public network there was a lot
of lag and jitter, probably due to clients having to redo all of the
inputs starting from the last one that the server confirms.

the clients no longer redo all the inputs so the game might not be
jittery any more, although i can't really test it out.