players can quit match, added more error messages
Clients can now open up the menu in the middle of the game to quit game,
or match.

Added some more error messages at different parts.

when the server is closed if any clients are connected they are informed
that they have been disconnected from the server.

if players are disconnected in a match they are shown an error message
and returned to the title screen.