Some sounds, changed drawing order, new control
some sound effects have been added to the game, so far all the gunshots
and reloading have been added, and all button sound effects.

the drawing order has been changed so the sniper's line of sight draws
on top of the foreground so it doesn't look weird.

if a large part of a player is under a bush then he doesn't get drawn on
screen that way they can actually hide.

players now have to right click the mouse button in order to pick up
gun.